+++ tests/fixtures/services/secrets.ts
@@ -1,4 +1,4 @@
-import { HasPermission, HookParams, Service, ServicePolicy, ServiceRequest, ServiceResponse, Before } from '../../../lib'
+import { HasPermission, HookParams, Service, ServicePolicy, ServiceRequest, ServiceResponse, Before, ServicePolicyDenial } from '../../../lib'
 
 class SecretsServicePolicy extends ServicePolicy {
   async index(req: ServiceRequest): Promise<boolean> {
@@ -7,6 +7,10 @@ class SecretsServicePolicy extends ServicePolicy {
 
   async post(req: ServiceRequest): Promise<boolean> {
     return this.ctx.state.key === 'abc123'
+  }
+
+  async put(req: ServiceRequest): Promise<boolean | ServicePolicyDenial> {
+    return new ServicePolicyDenial({ message: 'Method not implemented yet. Come back later' }, 405)
   }
 }
 
@@ -21,6 +25,13 @@ export default class SecretsService implements Service {
   @Before((hook: HookParams) => hook.req.ctx.state.key = hook.req.body.key)
   @HasPermission(SecretsServicePolicy, 'post')
   async post(req: ServiceRequest): Promise<ServiceResponse> {
+    return {
+      status: 204
+    }
+  }
+
+  @HasPermission(SecretsServicePolicy, 'put')
+  async put(req: ServiceRequest): Promise<ServiceResponse> {
     return {
       status: 204
     }